Create Modified Subarray Sum Problem
PrepBuddy is looking for a new problem statement for the students. She came across subarray sum problem, in subarray sum problem you are given an array and you have to find a subarray that has sum upto a given number. She thought this might be easy for the students so she modified it a bit. 
Now, the array 
A
 is concatenated 
K
 times to form a new array 
A
, i.e 
A
A
A
A
A
.
.
k
 times, now in this 
A
, you have to find the number of subarrays having a sum less than equal to a given number 
M
.
Since the answer can be very large you have to print the answer mod 
10
9
+
7
.