Create Median of Sorted Array
Sorting is another very important concept in competitive programming. Many of the problems can be solved easily if the elements are arranged in sorted order. We are going to solve one such problem. 
We all have studied the median in our schools, it is a simple concept. Let's say we have 
N
 numbers arranged in sorted order. Median is nothing but the mid element in case of an odd number of elements and in case of even number of elements it is the average of the two mid elements.

To keep your task simple, we have kept 
N
 as odd for you. So, your task is to find the median among 
N
 elements.