Create Arranging the Stones
Rahul and Tina have collected a few stones. They are given an array and are asked to arrange the stones in an array according to their weights. Tina is given the first half of the array and she arranges the weight of stones in increasing order and Rahul does the same task for his stones in 2nd half of the array.

Now PrepBuddy wants you to completely sort this array from start to end containing the weight of stones placed.