Create Sort the Numbers
Some problems are like a mirage. They seem to have a different solution when you see it first but as you deep dive you realize the solution is something else.

Next, we are going to solve one such problem.

You are given an array consisting of 0's, 1's and 2's. Your task is to sort the array.