Create Power Set of Strings
Many students face the issue with questions that are based on recursion and require string manipulation hence we have taken this as an agenda to make all of your very comfortable with such questions. Keeping this in mind we bring to you another such problem. The problem is to generate the power set of a given string in lexicographical order.

Power set 
P
(
S
)
 of a set 
S
 is the set of all subsets of 
S
. For example 
S
=
{
a
,
b
,
c
}
 then 
P
(
s
)
=
{
{
}
,
{
a
}
,
{
b
}
,
{
c
}
,
{
a
,
b
}
,
{
a
,
c
}
,
{
b
,
c
}
,
{
a
,
b
,
c
}
}
. Print without empty set.