Create Search Triplets
PrepBud has an array of integers. She was going through the numbers of the array and suddenly she realized there exist triplets in the array such that the sum of two elements equals the third element(not particularly in any order). She created a problem statement for you out of this situation. Given an array of distinct integers find out a triplet that satisfies the given condition.